pyproject.toml
[project]
name = "iceberg-viewer"
version = "0.1.0"
description = "Iceberg Navigator (icectl): a CLI for browsing Apache Iceberg catalogs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "click>=8.1",
  "pyyaml>=6.0",
  "tabulate>=0.9",
  # Planned/used for integration features
  "pyiceberg>=0.6; python_version >= '3.12'",
  "pyarrow>=15.0; python_version >= '3.12'",
  "pyiceberg>=0.9.1",
  "pyarrow>=21.0.0",
]

[project.scripts]
icectl = "icectl.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q -m 'not integration'"
pythonpath = ["src"]
markers = [
  "integration: tests requiring running docker-compose services",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=8.4.1",
    "ruff>=0.12.10",
]
